model sync with case processor

+++ src/main/java/uk/gov/ons/census/caseapisvc/exception/CaseReferenceNotFoundException.java
@@ -1,7 +1,7 @@
 package uk.gov.ons.census.caseapisvc.exception;
 
 public class CaseReferenceNotFoundException extends RuntimeException {
-  public CaseReferenceNotFoundException(Long caseId) {
+  public CaseReferenceNotFoundException(int caseId) {
     super(String.format("Case Reference '%s' not found", caseId));
   }
 }
