+++ src/test/java/uk/gov/ons/census/caseapisvc/endpoint/CaseEndpointIT.java
@@ -223,7 +223,7 @@ public class CaseEndpointIT {
   @Test
   public void shouldRetrieveACaseWithEventsWhenSearchingByCaseReference() throws Exception {
     Case expectedCase = createOneTestCaseWithEvent();
-    String expectedCaseRef = Long.toString(expectedCase.getCaseRef());
+    String expectedCaseRef = Integer.toString(expectedCase.getCaseRef());
 
     HttpResponse<JsonNode> response =
         Unirest.get(createUrl("http://localhost:%d/cases/ref/%s", port, expectedCaseRef))
@@ -242,7 +242,7 @@ public class CaseEndpointIT {
   @Test
   public void shouldRetrieveACaseWithoutEventsWhenSearchingByCaseReference() throws Exception {
     Case expectedCase = createOneTestCaseWithoutEvents();
-    String expectedCaseRef = Long.toString(expectedCase.getCaseRef());
+    String expectedCaseRef = Integer.toString(expectedCase.getCaseRef());
 
     HttpResponse<JsonNode> response =
         Unirest.get(createUrl("http://localhost:%d/cases/ref/%s", port, expectedCaseRef))
@@ -262,7 +262,7 @@ public class CaseEndpointIT {
   public void shouldRetrieveACaseWithoutEventsByDefaultWhenSearchingByCaseReference()
       throws Exception {
     Case expectedCase = createOneTestCaseWithoutEvents();
-    String expectedCaseRef = Long.toString(expectedCase.getCaseRef());
+    String expectedCaseRef = Integer.toString(expectedCase.getCaseRef());
 
     HttpResponse<JsonNode> response =
         Unirest.get(createUrl("http://localhost:%d/cases/ref/%s", port, expectedCaseRef))
