+++ src/test/java/uk/gov/ons/census/caseapisvc/utility/DataUtils.java
@@ -18,10 +18,10 @@ import uk.gov.ons.census.caseapisvc.model.entity.UacQidLink;
 public class DataUtils {
 
   private static final UUID TEST1_CASE_ID = UUID.fromString("2e083ab1-41f7-4dea-a3d9-77f48458b5ca");
-  private static final Long TEST1_CASE_REFERENCE_ID = 123L;
+  private static final Integer TEST1_CASE_REFERENCE_ID = 123;
 
   private static final UUID TEST2_CASE_ID = UUID.fromString("3e948f6a-00bb-466d-88a7-b0990a827b53");
-  private static final Long TEST2_CASE_REFERENCE_ID = 456L;
+  private static final Integer TEST2_CASE_REFERENCE_ID = 456;
 
   private static final String TEST_UPRN = "123";
 
@@ -41,7 +41,7 @@ public class DataUtils {
         createCase(TEST2_CASE_ID, TEST2_CASE_REFERENCE_ID));
   }
 
-  private static Case createCase(UUID caseId, Long caseRef) {
+  private static Case createCase(UUID caseId, Integer caseRef) {
     List<UacQidLink> uacQidLinks = new LinkedList<>();
     List<Event> events = new LinkedList<>();
